Keep child quick swap in navigation bar
